:sparkles: 파라미터값 없을 때 예외, 파일 하나로 합침
